Move Button common style into its folder path
